Refactor: moved css logic to css.py file, moved misc functions to utils.py
