+++ render.py
@@ -1,16 +1,14 @@
 import tkinter
-import tkinter.font
 from tkinter import ttk
-from browse import URL
-from browse import Text
-import browse as b
+from browse import URL, Text, HTMLParser, Element
+from css import CSSParser, style, cascade_priority
+from utils import tree_to_list, DrawRect, DrawText, get_font
 import sys
 
 WIDTH, HEIGHT = 800, 600
 HSTEP, VSTEP = 13, 18
 SCROLL_STEP = 100
 MAX_Y = 600
-FONTS = {}
 BLOCK_ELEMENTS = [
     "html", "body", "article", "section", "nav", "aside",
     "h1", "h2", "h3", "h4", "h5", "h6", "hgroup", "header",
@@ -19,117 +17,6 @@ BLOCK_ELEMENTS = [
     "figcaption", "main", "div", "table", "form", "fieldset",
     "legend", "details", "summary"
 ]
-INHERITED_PROPERTIES = {
-    "font-size": "16px",
-    "font-style": "normal",
-    "font-weight": "normal",
-    "color": "black",
-}
-
-class CSSParser:
-    def __init__(self, s):
-        self.s = s
-        self.i = 0 # Current pos.
-
-    def whitespace(self):
-        """
-        Whitespace increments have no parsed data to return.
-        """
-        while self.i < len(self.s) and self.s[self.i].isspace():
-            self.i += 1
-
-    def word(self):
-        """
-        Increments through word characters, returns substring it moved through.
-        """
-        start = self.i
-        while self.i < len(self.s):
-            if self.s[self.i].isalnum() or self.s[self.i] in "#-.%":
-                self.i += 1
-            else:
-                break
-        if not (self.i > start):
-            raise Exception("Parsing error")
-        return self.s[start:self.i]
-    
-    def literal(self, literal):
-        """
-        Checks for literals, punctutation chars.
-        """
-        if not (self.i < len(self.s) and self.s[self.i] == literal):
-            raise Exception("Parsing error")
-        self.i += 1
-
-    def pair(self):
-        prop = self.word()
-        self.whitespace()
-        self.literal(":")
-        self.whitespace()
-        val = self.word()
-        return prop.casefold(), val
-    
-    def body(self):
-        pairs = {}
-        while self.i < len(self.s) and self.s[self.i] != "}":
-            try:
-                prop, val = self.pair()
-                pairs[prop] = val
-                self.whitespace()
-                self.literal(";")
-                self.whitespace()
-            except Exception:
-                why = self.ignore_until([";", "}"])
-                if why == ";":
-                    self.literal(";")
-                    self.whitespace()
-                else:
-                    break
-        return pairs
-    
-    def parse(self):
-        rules = []
-        while self.i < len(self.s):
-            try:
-                self.whitespace()
-                selector = self.selector()
-                self.literal("{")
-                self.whitespace()
-                body = self.body()
-                self.literal("}")
-                rules.append((selector, body))
-            except Exception:
-                why = self.ignore_until(["}"])
-                if why == "}":
-                    self.literal("}")
-                    self.whitespace()
-                else:
-                    break
-        return rules
-    
-    def ignore_until(self, chars):
-        """
-        Error handling for property-value pairs that don't parse.
-        """
-        while self.i < len(self.s):
-            if self.s[self.i] in chars:
-                return self.s[self.i]
-            else:
-                self.i += 1
-        return None
-    
-    def selector(self):
-        """
-        Parsing function to create selector objects
-        """
-        out = TagSelector(self.word().casefold())
-        self.whitespace()
-        while self.i < len(self.s) and self.s[self.i] != "{":
-            tag = self.word()
-            descendant = TagSelector(tag.casefold())
-            out = DescendantSelector(out, descendant)
-            self.whitespace()
-        return out
-    
 DEFAULT_STYLE_SHEET = CSSParser(open("browser.css").read()).parse()
     
 class Browser:
@@ -163,13 +50,13 @@ class Browser:
         Obtains source code, delegates to other methods.
         """
         body = url.request()
-        self.nodes = b.HTMLParser(body).parse()
+        self.nodes = HTMLParser(body).parse()
         #b.print_tree(self.nodes) TO PRINT TREE IN TERMINAL
 
         rules = self.DEFAULT_STYLE_SHEET.copy()
         links = [node.attributes["href"]
             for node in tree_to_list(self.nodes, [])
-            if isinstance(node, b.Element)
+            if isinstance(node, Element)
             and node.tag == "link"
             and node.attributes.get("rel") == "stylesheet"
             and "href" in node.attributes]
@@ -290,7 +177,11 @@ class Browser:
         if e.width != WIDTH and e.height != HEIGHT:
             WIDTH = e.width
             HEIGHT = e.height
-            #paint_tree(self.document, self.display_list)
+            self.document = DocumentLayout(self.nodes)
+
+            self.document.layout()
+            self.display_list = []
+            paint_tree(self.document, self.display_list)
             self.draw()
 
 class DocumentLayout:
@@ -312,13 +203,10 @@ class DocumentLayout:
         self.x = HSTEP
         self.y = VSTEP
         child.layout()
-        #self.display_list = child.display_list
         self.height = child.height
-
 
     def paint(self):
         return []
-
 
 class BlockLayout:
     def __init__(self, node, parent, previous):
@@ -337,7 +225,6 @@ class BlockLayout:
         self.width = 0
         self.height = 0 
 
-
     def layout_mode(self):
         """
         Determines whether an element should be laid out via the recurse,
@@ -345,14 +232,13 @@ class BlockLayout:
         """
         if isinstance(self.node, Text):
             return "inline"
-        elif any([isinstance(child, b.Element) and \
+        elif any([isinstance(child, Element) and \
                   child.tag in BLOCK_ELEMENTS for child in self.node.children]):
             return "block"
         elif self.node.children:
             return "inline"
         else:
             return "block"
-
 
     def layout(self):        
         self.x = self.parent.x
@@ -382,38 +268,6 @@ class BlockLayout:
             self.recurse(self.node)
             self.flush()
             self.height = self.cursor_y
-    
-    def layout_intermediate(self):
-        previous = None
-        for child in self.node.children:
-            next = BlockLayout(child, self, previous)
-            self.children.append(next)
-            previous = next
-
-    def open_tag(self, tag):
-        if tag == "i":
-            self.style = "italic"
-        elif tag == "b":
-            self.weight = "bold"
-        elif tag == "small":
-            self.size -= 2
-        elif tag == "big":
-            self.size += 4
-        elif tag == "br":
-            self.flush()
-
-    def close_tag(self, tag):
-        if tag == "i":
-            self.style = "roman"
-        elif tag == "b":
-            self.weight = "normal"
-        elif tag == "small":
-            self.size += 2
-        elif tag == "big":
-            self.size -= 4
-        elif tag == "p":
-            self.flush()
-            self.cursor_y += VSTEP
 
     def recurse(self, node):
         if isinstance(node, Text):
@@ -436,8 +290,6 @@ class BlockLayout:
         size = int(float(node.style["font-size"][:-2]) * .75)
         font = get_font(size, weight, style)
         color = node.style["color"]
-
-        #font = get_font(self.size, self.weight, self.style)
         w = font.measure(word)
 
         if self.cursor_x + w > self.width:
@@ -485,130 +337,10 @@ class BlockLayout:
                 cmds.append(DrawText(x, y, word, font, color))
         return cmds
 
-    
-class DrawText:
-    def __init__(self, x1, y1, text, font, color):
-        self.top = y1
-        self.left = x1
-        self.text = text
-        self.font = font
-        self.color = color
-        self.bottom = y1 + font.metrics('linespace')
-
-    def execute(self, scroll, canvas):
-        canvas.create_text(
-            self.left, self.top - scroll,
-            text = self.text,
-            font = self.font,
-            fill = self.color,
-            anchor='nw')
-
-class DrawRect:
-    def __init__(self, x1, y1, x2, y2, color):
-        self.top = y1
-        self.left = x1
-        self.bottom = y2
-        self.right = x2
-        self.color = color
-
-    def execute(self, scroll, canvas):
-        canvas.create_rectangle(
-            self.left, self.top - scroll,
-            self.right, self.bottom - scroll,
-            width = 0,
-            fill = self.color
-        )
-    
-class TagSelector:
-    def __init__(self, tag):
-        self.tag = tag.lower()
-        self.priority = 1
-    
-    def matches(self, node):
-        return isinstance(node, b.Element) and self.tag == node.tag
-    
-
-class DescendantSelector:
-    def __init__(self, ancestor, descendant):
-        self.ancestor = ancestor
-        self.descendant = descendant
-        self.priority = ancestor.priority + descendant.priority
-
-    def matches(self, node):
-        if not self.descendant.matches(node): return False
-        while node.parent:
-            if self.ancestor.matches(node.parent): return True
-            node = node.parent
-        return False
-
-def style(node, rules):
-    node.style = {}
-
-    # Inheritance
-    for property, default_value in INHERITED_PROPERTIES.items():
-        if node.parent:
-            node.style[property] = node.parent.style[property]
-        else:
-            node.style[property] = default_value
-
-    # CSS Rules Applied
-    for selector, body in rules:
-        if not selector.matches(node): continue
-        for property, value in body.items():
-            node.style[property] = value
-        
-    # Inline styles
-    if isinstance(node, b.Element) and "style" in node.attributes:
-        pairs = CSSParser(node.attributes["style"]).body()
-        for property, value in pairs.items():
-            node.style[property] = value
-    
-    # Compute font sizes (%)
-    if node.style["font-size"].endswith("%"):
-        if node.parent:
-            parent_font_size = node.parent.style["font-size"]
-        else:
-            parent_font_size = INHERITED_PROPERTIES["font-size"]
-        node_pct = float(node.style["font-size"][:-1]) / 100
-        parent_px = float(parent_font_size[:-2])
-        node.style["font-size"] = str(node_pct * parent_px) + "px"
-
-    # Recurse children
-    for child in node.children:
-        style(child, rules)
-
-def cascade_priority(rule):
-    selector, body = rule
-    return selector.priority
-
-def get_font(size, weight, style):
-    """
-    Stores font if not in cache memory, otherwise stores font for future
-    reference. Impactful for Windows and Linux, not so much for MacOS.
-    """
-    key = (size, weight, style)
-    if key not in FONTS:
-        font = tkinter.font.Font(
-            size=size,
-            weight=weight,
-            slant=style
-            )
-        label = tkinter.Label(font=font)
-        FONTS[key] = (font, label)
-    return FONTS[key][0]
-
 def paint_tree(layout_object, display_list):
     display_list.extend(layout_object.paint())
-
     for child in layout_object.children:
         paint_tree(child, display_list)
-
-def tree_to_list(tree, list):
-    list.append(tree)
-    for child in tree.children:
-        tree_to_list(child, list)
-    return list
-
 
 def start(arg):
     Browser().load(URL(arg))
